+++ src/labs_examples/objects_classes_methods/labs/oop/C_blackjack2/models/Player.java
@@ -1,69 +1,83 @@
-package labs_examples.objects_classes_methods.labs.oop.C_blackjack2;
+package labs_examples.objects_classes_methods.labs.oop.C_blackjack2.models;
 
 import java.util.Scanner;
 
 public class Player {
-    private Hand hand = new Hand();
+
+    private final Hand hand = new Hand();
     private String name;
     private int potValue;
     private int bet;
 
-
-
-    public void newUser() {
-        Scanner scanner = new Scanner(System.in);
-        System.out.println("What's your name?");
-        this.name = scanner.next();
-
-        System.out.println("Hello " + this.name + "! How much money would you like to start with?");
-        if (scanner.hasNextInt()) {
-            this.potValue = scanner.nextInt();
-        }
+    public int getBet() {
+        return bet;
     }
+
+    public Hand getHand() {
+        hand.pause();
+        return hand;
+    }
+
+    //TODO - Jared edit - you have a name field, use it a bit more
+    public String getName() {
+        return name;
+    }
+
     public int getPotValue() {
         return potValue;
     }
 
-
-    public int setPotValue(int potValue) {
+    public void setPotValue(int potValue) {
         this.potValue = potValue;
-        return potValue;
     }
 
-
-    @Override
-    public String toString() {
-        return "Player{" +
-                "potValue=" + potValue +
-                '}';
+    public void newUser() {
+        Scanner scanner = new Scanner(System.in);
+        System.out.println("What's your name?");
+        name = scanner.next();
+        //TODO - Jared edit - this only necessary when a passed variable has same name as an instance variable
+        System.out.println("Hello " + name + "! How much money would you like to start with?");
+        if (scanner.hasNextInt()) {
+            potValue = scanner.nextInt();
+        }
     }
-
 
     public void handleBets() {
         Scanner scanner = new Scanner(System.in);
         System.out.println("\nHow much would you like to bet?");
         System.out.println("You currently have $" + potValue);
 
-        int bet = scanner.nextInt();
+        int proposedBet = scanner.nextInt();
+        //TODO - Jared edit - it is best for ambiguity to use separate names for members of methods vs members of class
+        // unless necessary, this also removes the need to use this
 
-        if (bet <= potValue) {
-
-            if (bet == potValue) {
+        if (proposedBet <= potValue) {
+            if (proposedBet == potValue) {
                 System.out.println("Are you sure you want to go all in? (y/n) ");
                 String userInput = scanner.next();
 
                 if (userInput.equalsIgnoreCase("y")) {
                     System.out.println("Nice! Good luck!");
                 } else {
-                    bet = getAdjustedBet();
+                    proposedBet = getAdjustedBet();
                 }
             }
-            this.bet = bet;
         } else {
-            this.bet = getAdjustedBet();
+            proposedBet = getAdjustedBet();
         }
+        bet = proposedBet;
     }
 
+    public int getAdjustedBet() {
+        int newBet = 0;
+        Scanner scanner = new Scanner(System.in);
+        do {
+            System.out.println("\nPlease bet less than $" + potValue);
+            System.out.println("\nHow much would you like to bet?");
+            newBet = scanner.nextInt();
+        } while (newBet > potValue);
+        return newBet;
+    }
 
     public void hitOrStay(Deck deck) {
         hand.pause();
@@ -94,33 +108,8 @@ public class Player {
                 System.out.println("Your total is " + hand.getHandValue());
                 break;
             }
-
         }
         while (hand.getHandValue() < 21);
-    }
-
-    public int getAdjustedBet() {
-        int newBet = 0;
-        Scanner scanner = new Scanner(System.in);
-        do {
-            System.out.println("\nPlease bet less than $" + potValue);
-            System.out.println("\nHow much would you like to bet?");
-            newBet = scanner.nextInt();
-        } while (newBet > potValue);
-        return newBet;
-    }
-
-    public int getBet() {
-        return bet;
-    }
-
-    public Hand getHand() {
-        hand.pause();
-        return hand;
-    }
-
-    public void setBet(int bet) {
-        this.bet = bet;
     }
 
     public void dealerHits(Deck deck) {
@@ -131,13 +120,18 @@ public class Player {
                 System.out.println("\nMy new card is " + hand.getCard(i + 2));
                 i++;
             } while (hand.getHandValue() <= 16);
-
         }
-
     }
 
-    public void printBalance(Player player) {
-        System.out.println("Your currently have $" + getPotValue() + " left");
+    public void printBalance() {
+        System.out.println("\nYour currently have $" + getPotValue() + " left");
+    }
+
+    @Override
+    public String toString() {
+        return "Player{" +
+                "potValue=" + potValue +
+                '}';
     }
 }
